# Conflicts:
#	fluXis/Graphics/Shaders/ShaderStackContainer.cs
#	fluXis/Map/Structures/Events/ShaderEvent.cs
#	fluXis/Scripting/Runners/StoryboardScriptRunner.cs
#	fluXis/Storyboards/Drawables/DrawableStoryboard.cs

+++ fluXis/Screens/Edit/EditorMap.cs
@@ -19,6 +19,7 @@ using fluXis.Storyboards;
 using fluXis.Utils;
 using fluXis.Utils.Extensions;
 using JetBrains.Annotations;
+using Midori.Utils;
 using Newtonsoft.Json;
 using osu.Framework.Graphics;
 using osu.Framework.Threading;
@@ -326,7 +327,7 @@ public class EditorMap : IVerifyContext
 
     private void throwMissingHandler(ITimedObject obj) => throw new ArgumentException($"Type '{obj.GetType().Name}' does not have a change handler associated with it.");
 
-    public void ApplyOffsetToAll(float offset) => notifiers.ForEach(n => n.ApplyOffset(offset));
+    public void ApplyOffsetToAll(double offset) => notifiers.ForEach(n => n.ApplyOffset(offset));
 
     public void Sort()
     {
@@ -386,7 +387,7 @@ public class EditorMap : IVerifyContext
         void Remove(ITimedObject obj);
         void Update(ITimedObject obj);
 
-        void ApplyOffset(float offset);
+        void ApplyOffset(double offset);
 
         bool Matches(Type type);
     }
@@ -444,7 +445,7 @@ public class EditorMap : IVerifyContext
             OnTypedUpdate?.Invoke((T)obj);
         }
 
-        public void ApplyOffset(float offset)
+        public void ApplyOffset(double offset)
         {
             foreach (var obj in list)
             {
